+++ src/test/java/io/hhplus/tdd/point/model/PointTest.java
@@ -4,9 +4,10 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import static io.hhplus.tdd.point.model.TransactionType.CHARGE;
+import static io.hhplus.tdd.point.model.TransactionType.USE;
 import static org.assertj.core.api.Assertions.assertThat;
 
-class ChargePointTest {
+class PointTest {
 
     @DisplayName("충전 포인트를 생성한다.")
     @Test
@@ -23,4 +24,18 @@ class ChargePointTest {
         assertThat(chargePoint.getTransactionType()).isEqualTo(CHARGE);
     }
 
+    @DisplayName("사용 포인트를 생성한다.")
+    @Test
+    void createUsePoint() {
+        // given
+        long value = 100_000L;
+        Amount amount = Amount.of(value);
+
+        // when
+        Point usePoint = UsePoint.of(amount);
+
+        // then
+        assertThat(usePoint.getAmount()).isEqualTo(-value);
+        assertThat(usePoint.getTransactionType()).isEqualTo(USE);
+    }
 }
